LMC 2700-Assignment 1
The goal of the site is to represent aspects of myself that I wish to share with others such as information about myself and my projects. It is also to reflect the minimalist impression I want to present to others. For the design aesthetic, I chose to keep the colors minimal to black and white to make the typography and the animated intro on my home page stand out and easier to read. The material I used for the site is a CSS template from free CSS.